Unit 1 lesson 4
adding sockets to place hats on hooks, duplicated sockets, added more hats, added socket to head, having trouble  aligning the socket to the main camera so I can see a tiny bit of the bill of the hat in my view so I know the overall alignment is correct. No matter how I adjust the socket alignment on the main camera, I can't see the hat in my view, also the hat looks off center when I look in the mirror.