Finished Art Challenge step 4
On the XR Rig, on the left hand, on the palette, drill down into the button sizes, small , medium and large are off. I think I'll make the large one bigger and the medium one way smaller. Also, on the medium pressed colors, the alpha (transparency) is set to 0 and it needs to be dragged all the way up. I couldn't see this color difference, but I watched a video and it was there.