Art Challenge step 5 solved
Turn slider on paintbrush from 2d to 3d sound, in Spatial Blend in the audio source.  This allows the sound to be louder when you are closer and also in one ear if the other ear is turned away from it. Also in the 3d sound settings, change the logarithmic rolloff to Linear Rolloff, change the min distance to 0 and the max to 1, this approximates a 50% drop in sound as you move the paint brush away from you.